Refactored TransitModal, turned the function to display the text into a hook
